backend/deal_histories/requirements.txt
Django==3.0
djangorestframework==3.14.0
pandas==2.0.3
psycopg2==2.9.6
python-dotenv==1.0.0
